implemented the code optimization.
