+++ src/pages/_app.js
@@ -11,9 +11,7 @@ import { useEffect, useState } from "react";
 
 
 export default function App({ Component, pageProps }) {
-
   const [loading, setLoading] = useState(true);
-
   useEffect(() => {
     setTimeout(() => {
       setLoading(false);
