+++ src/redux/action/auth.js
@@ -79,7 +79,6 @@ export const updateProfile = (data) => {
   return async (dispatch) => {
     try {
       const headers = createHeaders(); // Use the createHeaders function here
-
       const response = await axios.put(`${baseUrl}/professionals`,data,{ headers });
       return dispatch({
         type: Types.UPDATE_PROFILE,
@@ -98,7 +97,6 @@ export const registerClient = (data) => {
   return async (dispatch) => {
     try {
       const headers = createHeaders(); // Use the createHeaders function here
-
       const response = await axios.post(`${baseUrl}/client`, data, { headers });
       return dispatch({
         type: Types.CLIENT_DATA,
@@ -117,7 +115,6 @@ export const clientData = () => {
   return async (dispatch) => {
     try {
       const headers = createHeaders(); // Use the createHeaders function here
-
       const response = await axios.get(`${baseUrl}/client`, { headers });
       return dispatch({
         type: Types.CLIENT_DATA,
@@ -132,69 +129,10 @@ export const clientData = () => {
   };
 };
 
-export const WorkplaceDataCreation = (data) => {
-  return async (dispatch) => {
-    try {
-      const headers = createHeaders(); // Use the createHeaders function here
-
-      const response = await axios.post(`${baseUrl}/workplaces`,data,{headers});
-      return dispatch({
-        type: Types.WORKPLACE_DATA,
-        data: response?.data,
-      });
-    } catch (err) {
-      return dispatch({
-        type: Types.WORKPLACE_DATA_FAILURE,
-        data: err.response,
-      });
-    }
-  };
-};
-
-
-export const GetIndividualWorkplace = (id) => {
-  return async (dispatch) => {
-    try {
-      const headers = createHeaders(); // Use the createHeaders function here
-
-      const response = await axios.get(`${baseUrl}/workplaces/${id}`,{headers});
-      return dispatch({
-        type: Types.GET_INDIVIDUAL_WORKPLACE_SUCCES,
-        data: response?.data,
-      });
-    } catch (err) {
-      return dispatch({
-        type: Types.WORKPLACE_DATA_FAILURE,
-        data: err.response,
-      });
-    }
-  };
-};
-
-export const WorkplaceDataEdit = (data,id) => {
-  return async (dispatch) => {
-    try {
-      const headers = createHeaders(); // Use the createHeaders function here
-
-      const response = await axios.put(`${baseUrl}/workplaces/${id}`,data,{headers});
-      return dispatch({
-        type: Types.WORKPLACE_DATA,
-        data: response?.data,
-      });
-    } catch (err) {
-      return dispatch({
-        type: Types.WORKPLACE_DATA_FAILURE,
-        data: err.response,
-      });
-    }
-  };
-};
-
 export const getClientById = (clientId) => {
   return async (dispatch) => {
     try {
       const headers = createHeaders(); // Use the createHeaders function here
-
       const response = await axios.get(`${baseUrl}/client/${clientId}`, { headers });
       return dispatch({
         type: Types.CLIENT_BY_ID,
@@ -213,7 +151,6 @@ export const deleteClient = (clientId) => {
   return async (dispatch) => {
     try {
       const headers = createHeaders(); // Use the createHeaders function here
-
       const response = await axios.delete(`${baseUrl}/client/${clientId}`, { headers });
       return dispatch({
         type: Types.CLIENT_BY_ID,
@@ -228,59 +165,3 @@ export const deleteClient = (clientId) => {
   };
 };
 
-export const GetAllServices = () => {
-  return async (dispatch) => {
-    try {
-      const headers = createHeaders(); // Use the createHeaders function here
-
-      const response = await axios.get(`${baseUrl}/services`, { headers });
-      return dispatch({
-        type: Types.SERVICES,
-        data: response?.data,
-      });
-    } catch (err) {
-      return dispatch({
-        type: Types.SERVICES_FAILURE,
-        data: err.response,
-      });
-    }
-  };
-};
-
-export const GetIndividualService = (id) => {
-  return async (dispatch) => {
-    try {
-      const headers = createHeaders(); // Use the createHeaders function here
-
-      const response = await axios.get(`${baseUrl}/services/${id}`,{headers});
-      return dispatch({
-        type: Types.GET_INDIVIDUAL_SERVICE_SUCCESS,
-        data: response?.data,
-      });
-    } catch (err) {
-      return dispatch({
-        type: Types.SERVICES_FAILURE,
-        data: err.response,
-      });
-    }
-  };
-};
-
-export const ServiceDataEdit = (data,id) => {
-  return async (dispatch) => {
-    try {
-      const headers = createHeaders(); // Use the createHeaders function here
-
-      const response = await axios.put(`${baseUrl}/services/${id}`,data,{headers});
-      return dispatch({
-        type: Types.UPDATE_SERVICES,
-        data: response?.data,
-      });
-    } catch (err) {
-      return dispatch({
-        type: Types.UPDATE_SERVICES_FAILURE,
-        data: err.response,
-      });
-    }
-  };
-};
