+++ src/redux/action/profile.services.js
@@ -4,14 +4,9 @@ import axios from "axios";
 import createHeaders from "../../util/headers";
 
 export const CreateProfileService = (data) => {
-  console.log(
-    "🚀 ~ file: profile.services.js:7 ~ GetAllWorkplace ~ data:",
-    data
-  );
   return async (dispatch) => {
     try {
       const headers = createHeaders();
-
       const response = await axios.post(`${baseUrl}/services`, data, {
         headers,
       });
@@ -21,7 +16,64 @@ export const CreateProfileService = (data) => {
       });
     } catch (err) {
       return dispatch({
-        type: Types.WORKPLACE_DATA_FAILURE,
+        type: Types.CREATE_SERVICES_FAILURE,
+        data: err.response,
+      });
+    }
+  };
+};
+
+export const GetAllServices = () => {
+  return async (dispatch) => {
+    try {
+      const headers = createHeaders(); // Use the createHeaders function here
+      const response = await axios.get(`${baseUrl}/services`, { headers });
+      return dispatch({
+        type: Types.SERVICES,
+        payload: response?.data,
+      });
+    } catch (err) {
+      return dispatch({
+        type: Types.SERVICES_FAILURE,
+        data: err.response,
+      });
+    }
+  };
+};
+
+export const GetIndividualService = (id) => {
+  return async (dispatch) => {
+    try {
+      const headers = createHeaders(); // Use the createHeaders function here
+      const response = await axios.get(`${baseUrl}/services/${id}`,{headers});
+      return dispatch({
+        type: Types.GET_INDIVIDUAL_SERVICE_SUCCESS,
+        data: response?.data,
+      });
+    } catch (err) {
+      return dispatch({
+        type: Types.SERVICES_FAILURE,
+        data: err.response,
+      });
+    }
+  };
+};
+
+
+
+export const ServiceDataEdit = (data,id) => {
+  return async (dispatch) => {
+    try {
+      const headers = createHeaders(); // Use the createHeaders function here
+
+      const response = await axios.put(`${baseUrl}/services/${id}`,data,{headers});
+      return dispatch({
+        type: Types.UPDATE_SERVICES,
+        data: response?.data,
+      });
+    } catch (err) {
+      return dispatch({
+        type: Types.UPDATE_SERVICES_FAILURE,
         data: err.response,
       });
     }
