+++ src/redux/action/workplace.js
@@ -7,11 +7,65 @@ export const GetAllWorkplace = () => {
   return async (dispatch) => {
     try {
       const headers = createHeaders();
-
       const response = await axios.get(`${baseUrl}/workplaces`, { headers });
       return dispatch({
         type: Types.WORKPLACE_DATA,
         payload: response?.data,
+      });
+    } catch (err) {
+      return dispatch({
+        type: Types.WORKPLACE_DATA_FAILURE,
+        data: err.response,
+      });
+    }
+  };
+};
+
+export const WorkplaceDataCreation = (data) => {
+  return async (dispatch) => {
+    try {
+      const headers = createHeaders(); // Use the createHeaders function here
+      const response = await axios.post(`${baseUrl}/workplaces`,data,{headers});
+      return dispatch({
+        type: Types.WORKPLACE_DATA,
+        data: response?.data,
+      });
+    } catch (err) {
+      return dispatch({
+        type: Types.WORKPLACE_DATA_FAILURE,
+        data: err.response,
+      });
+    }
+  };
+};
+
+export const GetIndividualWorkplace = (id) => {
+  return async (dispatch) => {
+    try {
+      const headers = createHeaders(); // Use the createHeaders function here
+      const response = await axios.get(`${baseUrl}/workplaces/${id}`,{headers});
+      return dispatch({
+        type: Types.GET_INDIVIDUAL_WORKPLACE_SUCCES,
+        data: response?.data,
+      });
+    } catch (err) {
+      return dispatch({
+        type: Types.WORKPLACE_DATA_FAILURE,
+        data: err.response,
+      });
+    }
+  };
+};
+
+export const WorkplaceDataEdit = (data,id) => {
+  return async (dispatch) => {
+    try {
+      const headers = createHeaders(); // Use the createHeaders function here
+      const response = await axios.put(`${baseUrl}/workplaces/${id}`,data,{headers});
+      GetAllWorkplace()
+      return dispatch({
+        type: Types.WORKPLACE_DATA,
+        data: response?.data,
       });
     } catch (err) {
       return dispatch({
