+++ src/redux/constants/actionTypes.js
@@ -38,3 +38,4 @@ export const GET_INDIVIDUAL_SERVICE_SUCCESS = "GET_INDIVIDUAL_SERVICE_SUCCESS"
 export const UPDATE_SERVICES = "UPDATE_SERVICES"
 export const UPDATE_SERVICES_FAILURE = "UPDATE_SERVICES_FAILURE"
 export const CREATE_SERVICES = "CREATE_SERVICES"
+export const CREATE_SERVICES_FAILURE = "CREATE_SERVICES_FAILURE"
