+++ src/redux/reducer/workplace.js
@@ -3,6 +3,7 @@ import * as Types from "../constants/actionTypes";
 const initialState = {
   workplaceData: [],
   error: {},
+  individualWorkplace: null,
 };
 
 export const Workplace = (state = initialState, action) => {
@@ -14,8 +15,12 @@ export const Workplace = (state = initialState, action) => {
       };
     case Types.WORKPLACE_DATA_FAILURE:
       return { ...state, error: action?.payload?.response };
+    case Types.GET_INDIVIDUAL_WORKPLACE_SUCCES:
+      return {
+        ...state,
+        individualWorkplace: action.data,
+      };
     default:
       return state;
-      break;
   }
 };
