+++ src/redux/store.js
@@ -22,12 +22,12 @@ import reducer from "./reducer/rootReducer";
 import thunk from "redux-thunk";
 import { persistReducer, persistStore } from "redux-persist";
 import storage from "redux-persist/lib/storage";
+import { composeWithDevTools } from "redux-devtools-extension";
 const persistConfig = {
   key: "root",
   storage,
-  whitelist: ["Workplace"],
+  whitelist: ["Workplace","Services"],
 };
-import { composeWithDevTools } from "redux-devtools-extension";
 
 const persistedReducer = persistReducer(persistConfig, reducer);
